Made new Modal and remade Profile page
Made a new modal that tells user to configure their profile if their profile is not configured when the enter the home page. Also remade the profile page so that it works and looks better. In addition to that, made a new react component that makes other components show up on screen for "x" amount of time